requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-js-asset==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
drf-yasg==1.21.5
frozenlist==1.3.3
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.1
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
psycopg2-binary==2.9.6
PyJWT==2.6.0
pyotp==2.8.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
twilio==8.1.0
uritemplate==4.1.1
urllib3==1.26.15
yarl==1.8.2
zipp==3.15.0
